Struct poplar::channel::Channel

pub struct Channel<S, R>(/* private fields */)
where
    S: Serialize + DeserializeOwned,
    R: Serialize + DeserializeOwned;

Implementations

impl<S, R> Channel<S, R>
where S: Serialize + DeserializeOwned, R: Serialize + DeserializeOwned,

pub fn new_from_handle(handle: Handle) -> Channel<S, R>

pub fn create() -> Result<(Channel<S, R>, Handle), CreateChannelError>

Create a new channel. Returns one end as a Channel, and a Handle for the other end. Generally, the handle

pub fn subscribe_to_service( name: &str, ) -> Result<Channel<S, R>, SubscribeToServiceError>

pub fn send(&self, message: &S) -> Result<(), ChannelSendError>

pub fn try_receive(&self) -> Result<Option<R>, ChannelReceiveError>

Receive a message from the channel, if there’s one waiting. Returns Ok(None) if there are no pending messages to be received.

pub fn receive( &self, ) -> impl Future<Output = Result<R, ChannelReceiveError>> + '_

impl Channel<!, Handle>

pub fn register_service( name: &str, ) -> Result<Channel<!, Handle>, RegisterServiceError>