Function poplar::syscall::pci::pci_get_info

pub fn pci_get_info(
    buffer_ptr: *mut u8,
    buffer_size: usize,
) -> Result<usize, PciGetInfoError>

Makes a raw pci_get_info system call, given a pointer to a buffer and the size of the buffer. On success, returns the number of entries written into the buffer. For a nicer interface to this system call, see crate::ddk::pci::pci_get_info_slice or crate::ddk::pci::pci_get_info_vec - these are part of the DDK to avoid pulling the pci_types crate into everything that uses this crate.